requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
async-timeout==4.0.2
asyncpg==0.28.0
backcall==0.2.0
beautifulsoup4==4.12.0
boto3==1.26.119
botocore==1.29.119
bs4==0.0.1
cassandra-driver==3.25.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Cython==0.29.34
databases==0.8.0
decorator==5.1.1
distlib==0.3.4
dnspython==2.3.0
docx2txt==0.8
exceptiongroup==1.2.1
fastapi==0.103.2
filelock==3.7.1
future==0.18.3
FXrays==1.3.5
geomet==0.2.1.post1
idna==3.4
importlib-metadata==4.11.4
install==1.3.5
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
knot-floer-homology==1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
networkx==2.6.3
numpy==1.21.6
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
plink==2.4.1
prompt-toolkit==3.0.38
psycopg2==2.9.9
ptyprocess==0.7.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
pymemcache==4.0.0
pymongo==4.3.3
pypandoc==1.11
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==0.21.1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.2
s3transfer==0.6.0
sentry-sdk==1.16.0
six==1.16.0
snappy-manifolds==1.1.2
sniffio==1.3.1
soupsieve==2.4
spherogram==2.1
SQLAlchemy==1.4.52
starlette==0.27.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.14
virtualenv==20.14.1
wcwidth==0.2.6
zipp==3.8.0
